broke out posts and added to misc
