+++ ezomero/__init__.py
@@ -1,10 +1,4 @@
-from ._ezomero import (post_dataset,
-                       post_image,
-                       post_map_annotation,
-                       post_file_annotation,
-                       post_project,
-                       post_roi,
-                       get_image,
+from ._ezomero import (get_image,
                        get_image_ids,
                        get_map_annotation_ids,
                        get_map_annotation,
@@ -14,16 +8,22 @@ from ._ezomero import (post_dataset,
                        get_user_id,
                        get_original_filepaths,
                        put_map_annotation,
-                       print_map_annotation,
-                       print_groups,
-                       print_projects,
-                       print_datasets,
                        connect,
                        store_connection_params,
                        set_group)
 from ._misc import (filter_by_filename,
                     link_images_to_dataset,
-                    link_datasets_to_project)
+                    link_datasets_to_project,
+                    print_map_annotation,
+                    print_groups,
+                    print_projects,
+                    print_datasets)
+from ._posts import (post_dataset,
+                     post_image,
+                     post_map_annotation,
+                     post_file_annotation,
+                     post_project,
+                     post_roi)
 
 __all__ = ['post_dataset',
            'post_image',
