+++ ezomero/_misc.py
@@ -111,3 +111,82 @@ def _get_current_user(conn):
     if userid is None:
         userid = conn.getUserId()
     return userid
+
+
+# prints
+def print_map_annotation(conn, map_ann_id):
+    """Print some information and value of a map annotation.
+
+    Parameters
+    ----------
+    conn : ``omero.gateway.BlitzGateway`` object
+        OMERO connection.
+    map_ann_id : int
+        Id of the MapAnnotation to be displayed.
+    """
+    map_ann = conn.getObject('MapAnnotation', map_ann_id)
+    print(f'Map Annotation: {map_ann_id}')
+    print(f'Namespace: {map_ann.getNs()}')
+    print('Key-Value Pairs:')
+    for k, v in map_ann.getValue():
+        print(f'\t{k}:\t{v}')
+
+
+def print_groups(conn):
+    """Print all Groups with IDs and membership info.
+
+    Parameters
+    ----------
+    conn : ``omero.gateway.BlitzGateway`` object
+        OMERO connection.
+    """
+    user_id = conn.getUser().getId()
+    print("Groups:")
+    for g in conn.listGroups():
+        if g.getId() not in [0, 1, 2]:
+            owners, members = g.groupSummary()
+            owner_ids = [e.getId() for e in owners]
+            member_ids = [e.getId() for e in members]
+            if user_id in owner_ids:
+                group_status = 'owner'
+            elif user_id in member_ids:
+                group_status = 'member'
+            else:
+                group_status = ''
+            print(f'{g.getName():>25}: {g.getId()}\t{group_status}')
+
+
+def print_projects(conn):
+    """Print all available Projects.
+
+    Parameters
+    ----------
+    conn : ``omero.gateway.BlitzGateway`` object
+        OMERO connection.
+    """
+    print("Projects:")
+    for p in conn.listProjects():
+        print(f'\t{p.getName()}:\t{p.getId()}')
+
+
+def print_datasets(conn, project=None):
+    """Print all available Datasets for a given Project.
+
+    Parameters
+    ----------
+    conn : ``omero.gateway.BlitzGateway`` object
+        OMERO connection.
+    project : int or `None`, optional
+        ID of Project for which to list datasets. If project is `None`,
+        orphans are listed.
+    """
+    if project is not None:
+        p = conn.getObject("Project", project)
+        datasets = p.listChildren()
+        print(f'Datasets in Project \"{p.getName()}\":')
+    else:
+        datasets = conn.listOrphans("Dataset")
+        print('Orphaned Datsets:')
+
+    for d in datasets:
+        print(f"\t{d.getName()}:\t{d.getId()}")
